Move module `prove_rpc` from `ktool` to `proof`
